requirements.txt
altgraph==0.17
Appium-Python-Client==1.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.8
asyncio==3.4.3
attrs==23.1.0
autobahn==23.1.2
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
baidu-aip==2.2.18.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bleach==5.0.1
certifi==2020.12.5
cffi==1.14.5
channels==4.0.0
chardet==4.0.0
click==8.0.4
colorama==0.4.4
constantly==15.1.0
contourpy==1.1.1
cryptography==41.0.7
cssselect==1.1.0
cycler==0.12.1
daphne==4.0.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
dictop==0.2.1
Django==4.0.2
django-admin-global-sidebar==0.2.0
django-cors-headers==3.13.0
django-filter==22.1
django-simpleui==2022.3.15
django-static-fontawesome==5.14.0.0
django-static-jquery3==5.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.0.0
f5-icontrol-rest==1.3.13
f5-sdk==3.0.21
fastjsonschema==2.16.1
fonttools==4.45.1
future==0.18.2
gevent==23.9.1
greenlet==3.0.1
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
incremental==22.10.0
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.1
Jinja2==3.1.1
jmespath==0.10.0
jsonschema==4.14.0
junos-eznc==2.6.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kiwisolver==1.4.5
lxml==4.6.2
magic-import==0.1.3
MarkupSafe==2.1.1
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==2.0.4
moviepy==1.0.3
mysqlclient==2.1.0
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
ncclient==0.6.9
nest-asyncio==1.5.5
netaddr==0.8.0
netmiko==3.4.0
notebook==6.4.12
ntc-templates==3.0.0
numpy==1.24.4
openpyxl==3.0.7
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
paramiko==2.9.2
parsel==1.6.0
parso==0.8.3
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.0.1
pkgutil_resolve_name==1.3.10
proglog==0.1.9
prometheus-client==0.14.1
prompt-toolkit==3.0.30
Protego==0.1.16
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyeapi==0.8.4
PyExecJS==1.5.1
pygame==2.0.1
Pygments==2.13.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyntc==0.19.0
pynxos==0.0.5
pyOpenSSL==23.3.0
pyparsing==3.0.8
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.7
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
queuelib==1.5.0
requests==2.25.1
scp==0.13.6
Scrapy==2.4.1
selenium==3.141.0
Send2Trash==1.8.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2
spark-parser==1.8.9
sqlparse==0.4.2
stack-data==0.5.0
tenacity==8.0.1
terminado==0.15.0
textfsm==1.1.2
tinycss2==1.1.1
tornado==6.2
tqdm==4.54.1
traitlets==5.3.0
transitions==0.8.11
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uncompyle6==3.8.0
urllib3==1.26.2
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.2
xdis==6.0.3
xlrd==2.0.1
yamlordereddictloader==0.4.0
zipp==3.8.1
zope.event==5.0
zope.interface==5.2.0
